Start implementing kiosk mode
This is the display that will be shown on screens at the entrance of
event to allow attendees to sign in and out in a self-serve fashion.

The kiosk page can be accessed either by a kiosk short URL or by
clicking a link from the signed-in user page.

TODO: In the latter case, we also need to ensure that the user is signed
out before showing the kiosk mode screen.